requirements.txt
fastapi
langchain
fitz
google.generativeai
chromadb  
langchain_community 
langchain_google_genai
tools
frontend
sentence-transformers
PyMuPDF